CZERTAINLY-Core #489 Implement assigning approval profile to RA profile - endpoints transferred to RAProfileManagementController
